We should be as forward looking as possible
Right now the text only defines that future edns options must be
supported, but (IMHO) we should support all future versions of changes
to the DNS in general, not just future EDNS extensions.  So I proposed
text that rewords a bit to state this and uses EDNS as an example.